hostconf/app/requirements.txt
flask == 2.3.2
gunicorn == 20.1.0
pyyaml == 6.0
schema == 0.7.5
